Major restructuring of "DAWIntegration" section, DAW/plugin mode fully documented
Major revision of the section discussing plugdata within the DAW.

I created a new full summary page for plugdata in the DAW called plugdata-in-the-DAW ( prefer this over DAW integration, as it sounds more like we are integrating a DAW into plugdata, instead of integrating plugdata into a DAW - which I feel more accurately describes what we are doing). This document pulls the DAW information from the main branch Wiki, plus revises all the DAW integration documentation that was currently in the folder.

I broke down the document that discussed plugdata inside each DAW into their own sepaerate pages. This just seemed to make the most sense to me. **I updated the links found in SUMMARY.md to link to these pages now**. The main plugdata-in-the-DAW document links to all of these pages as well.